refactor: extract PositionShiftVisitor into its own file
Move PositionShiftVisitor out of IncrementalParsingService into a
separate file, following the project convention of one class per file.
